refactor(suite-native): split account selectors
