+++ suite-native/accounts/src/index.ts
@@ -13,7 +13,15 @@ export * from './components/AccountLabelFieldHint';
 export * from './components/AccountDetailsCard';
 export * from './components/TokenReceiveCard';
 export * from './hooks/useAccountLabelForm';
-export * from './selectors';
+export * from './selectors/common';
+export * from './selectors/getAccountListSections';
+export * from './selectors/selectAccountFiatBalance';
+export * from './selectors/selectAccountListSections';
+export * from './selectors/selectAccountTokenFiatBalance';
+export * from './selectors/selectFilteredDeviceAccountsGroupedByNetworkAccountType';
+export * from './selectors/selectFreshAccountAddress';
+export * from './selectors/selectHasDeviceAnySendAvailableAccount';
+export * from './selectors/selectVisibleAccountsWithLabel';
 export * from './hooks/useAccountAlerts';
 export * from './utils';
 export { type OnSelectAccount } from './types';
